Only send updates relevant to seurity app to security app

+++ app/services/frontend_updates.rb
@@ -9,25 +9,25 @@ class FrontendUpdates
       created_shifts.each_value do |shift_data|
         shift = shift_data.fetch(:shift)
         if shift.rota_published?
-          security_app_update_service.create_shift(shift: shift)
+          security_app_update_service.create_shift(shift: shift) if shift.staff_member.security?
         end
       end
       shift_updates.each_value do |shift_data|
         shift = shift_data.fetch(:shift)
         if shift.rota_published?
-          security_app_update_service.update_shift(shift: shift)
+          security_app_update_service.update_shift(shift: shift) if shift.staff_member.security?
         end
       end
       shift_deletes.each_value do |shift_data|
         shift = shift_data.fetch(:shift)
         if shift.rota_published?
-          security_app_update_service.delete_shift(shift: shift)
+          security_app_update_service.delete_shift(shift: shift) if shift.staff_member.security?
         end
       end
       staff_member_profile_updates.each_value do |update_data|
         staff_member = update_data.fetch(:staff_member)
         security_app_update_service
-          .update_staff_member_profile(staff_member: staff_member)
+          .update_staff_member_profile(staff_member: staff_member) if staff_member.security?
       end
       security_app_update_service.call
     rescue Exception => e
